Refactoring
Rename Struct to Structure
Rename Union to Dynamic

Rework KeyManager
Rework Structure
Rework Dynamic

Remove deprecated
Remove Type namespace
